Loadtxt.py init + Upload XSS.txt dataset
